Main/requirements.txt
django
opencv-python
tensorflow
pillow